DTS 版本2，id生成器配置存库，需要用http的类改为scala 添加java/scala json解析工具，单元测试及代码局部调整 Change-Id: I1b27cd6e97046cc1f007465b02e622c3a526e4da

+++ dts/dts-client/src/main/java/com/itiancai/galaxy/dts/annotation/Activity.java
@@ -26,5 +26,5 @@ public @interface Activity {
      * 主线程超时时间 单位:毫秒
      * @return
      */
-    long timeOut() default 30000;
+    int timeOut() default 30000;
 }
